Abstracted all routes to declutter server.js
